refactor:랜덤 팩 함수 분리
